// dist/Turkish/TurkishDependencyType.d.ts
/**
 * Enumerator class for Turkish Dependency tags
 */
export declare enum TurkishDependencyType {
    VOCATIVE = 0,
    SUBJECT = 1,
    DATIVE_ADJUNCT = 2,
    OBJECT = 3,
    POSSESSOR = 4,
    MODIFIER = 5,
    S_MODIFIER = 6,
    ABLATIVE_ADJUNCT = 7,
    DETERMINER = 8,
    SENTENCE = 9,
    CLASSIFIER = 10,
    LOCATIVE_ADJUNCT = 11,
    COORDINATION = 12,
    QUESTION_PARTICLE = 13,
    INTENSIFIER = 14,
    INSTRUMENTAL_ADJUNCT = 15,
    RELATIVIZER = 16,
    NEGATIVE_PARTICLE = 17,
    ETOL = 18,
    COLLOCATION = 19,
    FOCUS_PARTICLE = 20,
    EQU_ADJUNCT = 21,
    APPOSITION = 22
}